Moved test to the correct file
